pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "text_summarization_project"
version = "0.1.0"
description = "Text summarization and paraphrasing project"
authors = [
    { name = "lavanya", email = "lavanyachowdam22.4211@gmail.com" }
]
dependencies = [
    "transformers",
    "torch",
    "nltk",
    "networkx",
    "pyyaml",
    "numpy",
    "sentencepiece",
    "streamlit",
    "protobuf",
    "python-dotenv",
    "requests"
]